test: reorganize statistics tests
