+++ mathFunctions/statisticsFunctions/calculateAverage.ts
@@ -1,0 +1,1 @@
+export { calculateAverage } from '../../arrayFunctions/calculateAverage';
