+++ mathFunctions/statisticsFunctions/calculateMedian.ts
@@ -1,0 +1,1 @@
+export { calculateMedian } from '../../arrayFunctions/calculateMedian';
