+++ mathFunctions/statisticsFunctions/calculateMode.ts
@@ -1,0 +1,1 @@
+export { calculateMode } from '../../arrayFunctions/calculateMode';
