+++ mathFunctions/statisticsFunctions/calculateRange.ts
@@ -1,0 +1,1 @@
+export { calculateRange } from '../../arrayFunctions/calculateRange';
