+++ mathFunctions/statisticsFunctions/calculateStandardDeviation.ts
@@ -1,0 +1,1 @@
+export { calculateStandardDeviation } from '../../arrayFunctions/calculateStandardDeviation';
