+++ mathFunctions/statisticsFunctions/calculateVariance.ts
@@ -1,0 +1,1 @@
+export { calculateVariance } from '../../arrayFunctions/calculateVariance';
